Parse SELECT statements to AST
I've cheated a little bit here: I'm using "," to explode the list into
items rather than consume it token by token. As a result we're not
really checking for, eg. a proper terminator.

Still, this will probably do for the purposes of creating a simple AST
that we can then convert into JSON and return.