Use native types for literals
I was trying to be fancy by having nodes with a type (eg. BOOLEAN,
INTEGER) and a value. But thinking about it, I'm not sure this adds
anything other than complexity: the values already *have* types, so
there's no need to separately declare them.

Let's get rid of them for now. The only awkwardness this leaves is that
SELECT statements are still represented as a node. Maybe this is ok?
Or maybe we can do something better here too?