Parse statements recursively
Previously this was a bit hacky. Instead, let's try following a process
where we consume tokens one at a time from the statement until there's
nothing left.

This is done via a little class that extends String (is that eccentric?)
with a `consume` method.

Since `parseLiteral` now actually removes each literal from the string
as it parses it, `parseList` can call itself recursively to build an
array of values from a comma-separated string.

The main thing I'm not so keen on here is having to explicitly make sure
each matching pattern has a `^` (ie. match from the start). I couldn't
find a nice way to interpolate one Regex into another in Javascript,
and since we need a Regex for matching integers, we'll also need to
specify, eg. `/^TRUE/` and `/^FALSE/` - otherwise we're not matching
the start of the string and it all goes wonky...