Allow columns to be aliased
To get aliases I'm matching a more complex regex - "AS " followed by
an allowable column name. Is this cheating? I'm not sure. I'd originally
thought of parsing the whole statement into tokens, and then building
the AST from the tokens (that might still be worth investigating), but
for now this seems to work...